Extract engine metadata operations
read_metadata and write_metadata are a compact paired operation family around ffprobe tags and FFmpeg metadata writes. This moves them into engine_metadata.py while re-exporting both from mcp_video.engine for compatibility.

Constraint: Preserve tag extraction precedence, metadata validation, copy-codec write behavior, and result shapes.

Rejected: Move compare_quality with metadata | quality comparison is an analysis workflow with separate FFmpeg metrics behavior.

Confidence: high

Scope-risk: narrow

Directive: Keep paired operations together when they share a narrow contract and tests.

Tested: ruff check mcp_video/engine.py mcp_video/engine_metadata.py

Tested: ruff format --check mcp_video/engine.py mcp_video/engine_metadata.py

Tested: /opt/homebrew/bin/python3 read_metadata/write_metadata re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine_advanced.py -k 'metadata' tests/test_cli.py -k 'metadata' tests/test_server.py -k 'metadata' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
